Moved code related to displaying into a new class
